Moved lib and shared to src/ folder. Related fixes.
