+++ package.json
@@ -2,16 +2,16 @@
   "name": "cumulocity-cypress",
   "version": "0.2.4",
   "description": "Cypress commands for Cumulocity IoT",
-  "main": "lib/commands/index.js",
+  "main": "src/lib/commands/index.js",
   "scripts": {
-    "start-server": "http-server test/cypress/app/ --port 8080 --silent --mimetypes test/custom_mime_types.types --ext json",
+    "start-server": "http-server test/cypress/app/ --silent --port 8080 --mimetypes test/custom_mime_types.types --ext json",
     "test": "cross-env TZ=Europe/Berlin start-server-and-test start-server http://localhost:8080 test:run",
-    "test:run": "cross-env TZ=Europe/Berlin cypress run  --project test --browser chrome --config video=false",
+    "test:run": "cross-env TZ=Europe/Berlin cypress run --project test --browser chrome --config video=false",
     "test:open": "cross-env TZ=Europe/Berlin start-server-and-test start-server http://localhost:8080 debug",
     "debug": "cross-env CYPRESS_REMOTE_DEBUGGING_PORT=9222 cypress open --project test",
     "clean": "rimraf dist/ && rimraf packages/pactrunner/dist/ && rimraf packages/pactrunner/.yalc && rimraf packages/pactrunner/yalc.lock",
-    "copy-files": "copyfiles lib/**/*.*js lib/**/*.d.ts README.md package.json dist/",
-    "build": "npm run clean && npm run copy-files && tsc -b -v lib/ && cd dist/ && npm run clean-package-json",
+    "copy-files": "copyfiles --up 1 src/lib/**/*.*js src/lib/*.*js src/lib/**/*.d.ts dist/ && copyfiles README.md package.json dist/",
+    "build": "npm run clean && npm run copy-files && tsc -b -v src/lib/ && cd dist/ && npm run clean-package-json",
     "runner:yalc": "npm run clean && npm run build && cd dist/ && yalc publish && cd ../packages/pactrunner && yalc add cumulocity-cypress",
     "clean-package-json": "npm pkg delete 'devDependencies' && npm pkg delete 'scripts'"
   },
