+++ packages/pactrunner/cypress/support/e2e.ts
@@ -1,4 +1,4 @@
-import "cumulocity-cypress";
+import "cumulocity-cypress/lib/commands/";
 import "cumulocity-cypress/lib/commands/c8ypact";
 import "cumulocity-cypress/lib/commands/request";
