+++ src/lib/commands/common.ts
