+++ src/lib/commands/general.d.ts
