+++ src/lib/commands/index.ts
