+++ src/lib/commands/intercept.ts
