+++ src/lib/commands/login.ts
@@ -1,7 +1,7 @@
 const { _ } = Cypress;
 const semver = require("semver");
 import { ICredentials } from "@c8y/client";
-const { getAuthOptions, isAuth, resetClient } = require("./utils");
+const { getAuthOptions, isAuth, resetClient } = require("./../utils");
 
 declare global {
   namespace Cypress {
