+++ src/lib/locale/de.d.ts
