+++ src/lib/locale/de.mjs
