+++ src/lib/locale/en-GB.d.ts
