+++ src/lib/locale/en.d.ts
