+++ src/lib/tsconfig.json
@@ -8,7 +8,7 @@
     "lib": ["dom", "dom.iterable", "ES2020"],
     "types": ["cypress", "node"],
     "declaration": true,
-    "outDir": "../dist/lib",
+    "outDir": "../../dist/lib",
     "strict": true,
     "module": "Node16",
     "skipLibCheck": true,
