+++ src/shared/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "rootDir": ".",
-    "outDir": "../dist/shared",
+    "outDir": "../../dist/shared",
     "target": "ES2020",
     "lib": ["dom", "dom.iterable", "ES2020"],
     "types": ["node"],
